Extract shared atomic_install into fs_util
The report writer and the XMP sidecar writer each kept a near-identical
copy of the EXDEV-safe rename helper introduced by R-6. Consolidate
into a single fs_util module; report.rs and download/metadata.rs now
share one implementation and one set of tests.
