+++ src/fs_util.rs
@@ -1,0 +1,69 @@
+//! Shared filesystem primitives.
+
+use std::path::Path;
+
+/// Install `src` at `dst` atomically. Prefers `rename` (atomic on the same
+/// device); on EXDEV, copies to a sibling of `dst` on the destination device
+/// and renames that sibling into place so a mid-copy crash can't expose a
+/// half-written `dst`.
+pub(crate) fn atomic_install(src: &Path, dst: &Path) -> std::io::Result<()> {
+    if let Err(rename_err) = std::fs::rename(src, dst) {
+        let ext = dst.extension().and_then(|e| e.to_str()).unwrap_or("tmp");
+        let dst_sibling = dst.with_extension(format!("{ext}.kei-xdev-tmp-{}", std::process::id()));
+        if let Err(copy_err) = std::fs::copy(src, &dst_sibling) {
+            let _ = std::fs::remove_file(src);
+            tracing::warn!(
+                src = %src.display(),
+                dst = %dst.display(),
+                rename_err = %rename_err,
+                copy_err = %copy_err,
+                "rename failed and cross-device copy also failed"
+            );
+            return Err(rename_err);
+        }
+        if let Err(final_err) = std::fs::rename(&dst_sibling, dst) {
+            let _ = std::fs::remove_file(&dst_sibling);
+            let _ = std::fs::remove_file(src);
+            return Err(final_err);
+        }
+        let _ = std::fs::remove_file(src);
+    }
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn same_device_rename_succeeds() {
+        let dir = tempfile::tempdir().unwrap();
+        let src = dir.path().join("src.tmp");
+        let dst = dir.path().join("dst.json");
+        std::fs::write(&src, b"hello").unwrap();
+
+        atomic_install(&src, &dst).expect("atomic_install");
+
+        assert!(!src.exists(), "src must be consumed by the rename");
+        assert_eq!(std::fs::read(&dst).unwrap(), b"hello");
+
+        for entry in std::fs::read_dir(dir.path()).unwrap().flatten() {
+            let name = entry.file_name();
+            let name = name.to_string_lossy();
+            assert!(
+                !name.contains("kei-xdev-tmp"),
+                "unexpected sidecar tmp {name}"
+            );
+        }
+    }
+
+    #[test]
+    fn missing_src_returns_err_without_touching_dst() {
+        let dir = tempfile::tempdir().unwrap();
+        let src = dir.path().join("nope.tmp");
+        let dst = dir.path().join("dst.json");
+
+        assert!(atomic_install(&src, &dst).is_err());
+        assert!(!dst.exists(), "dst must not be created on failure");
+    }
+}
